+++ src/common/init.c
@@ -18,184 +18,13 @@
     #include <signal.h>
 #endif
 
-#define FF_ENSURE_ONLY_ONCE_IN_LIST(list, element) \
-    if(ffListFirstIndexComp(list, element, (bool(*)(const void*, const void*))ffStrbufEqual) < list->length - 1) \
-    { \
-        ffStrbufDestroy(ffListGet(list, list->length - 1)); \
-        --list->length; \
-    }
-
-static void pathsAddHome(FFlist* dirs, FFstate* state, const char* suffix)
-{
-    FFstrbuf* buffer = (FFstrbuf*) ffListAdd(dirs);
-    ffStrbufInitA(buffer, 64);
-    ffStrbufAppendS(buffer, state->passwd->pw_dir);
-    ffStrbufAppendS(buffer, suffix);
-    ffStrbufEnsureEndsWithC(buffer, '/');
-    FF_ENSURE_ONLY_ONCE_IN_LIST(dirs, buffer);
-}
-
-static void pathsAddAbsolute(FFlist* dirs, const char* path)
-{
-    FFstrbuf* buffer = (FFstrbuf*) ffListAdd(dirs);
-    ffStrbufInitA(buffer, 64);
-    ffStrbufAppendS(buffer, path);
-    ffStrbufEnsureEndsWithC(buffer, '/');
-    FF_ENSURE_ONLY_ONCE_IN_LIST(dirs, buffer);
-}
-
-static void pathsAddEnv(FFlist* dirs, const char* env)
-{
-    const char* envValue = getenv(env);
-    if(!ffStrSet(envValue))
-        return;
-
-    FFstrbuf value;
-    ffStrbufInitA(&value, 64);
-    ffStrbufAppendS(&value, envValue);
-
-    uint32_t startIndex = 0;
-    while (startIndex < value.length)
-    {
-        uint32_t colonIndex = ffStrbufNextIndexC(&value, startIndex, ':');
-        value.chars[colonIndex] = '\0';
-
-        if(!ffStrSet(value.chars + startIndex))
-        {
-            startIndex = colonIndex + 1;
-            continue;
-        }
-
-        pathsAddAbsolute(dirs, value.chars + startIndex);
-
-        startIndex = colonIndex + 1;
-    }
-
-    ffStrbufDestroy(&value);
-}
-
-#ifdef _WIN32
-static void pathsAddKnownFolder(FFlist* dirs, REFKNOWNFOLDERID folderId)
-{
-    PWSTR pPath;
-    if(SUCCEEDED(SHGetKnownFolderPath(folderId, 0, NULL, &pPath)))
-    {
-        FFstrbuf* buffer = (FFstrbuf*) ffListAdd(dirs);
-        ffStrbufInit(buffer);
-        ffStrbufSetWS(buffer, pPath);
-        ffStrbufReplaceAllC(buffer, '\\', '/');
-        ffStrbufEnsureEndsWithC(buffer, '/');
-        FF_ENSURE_ONLY_ONCE_IN_LIST(dirs, buffer);
-    }
-    CoTaskMemFree(pPath);
-}
-
-static void pathsAddEnvSuffix(FFlist* dirs, const char* env, const char* suffix)
-{
-    const char* value = getenv(env);
-    if(!ffStrSet(value))
-        return;
-
-    FFstrbuf* buffer = ffListAdd(dirs);
-    ffStrbufInitA(buffer, 64);
-    ffStrbufAppendS(buffer, value);
-    ffStrbufReplaceAllC(buffer, '\\', '/');
-    ffStrbufEnsureEndsWithC(buffer, '/');
-    ffStrbufAppendS(buffer, suffix);
-    ffStrbufEnsureEndsWithC(buffer, '/');
-    FF_ENSURE_ONLY_ONCE_IN_LIST(dirs, buffer);
-}
-#endif
-
-static void initConfigDirs(FFstate* state)
-{
-    ffListInit(&state->configDirs, sizeof(FFstrbuf));
-
-    pathsAddEnv(&state->configDirs, "XDG_CONFIG_HOME"); // On systems where this is not set (Windows & Apple presumably), this wil do nothing
-    pathsAddHome(&state->configDirs, state, "/.config/");
-
-    #ifdef _WIN32
-        pathsAddKnownFolder(&state->configDirs, &FOLDERID_RoamingAppData);
-        pathsAddKnownFolder(&state->configDirs, &FOLDERID_LocalAppData);
-    #elif defined(__APPLE__)
-        pathsAddHome(&state->configDirs, state, "/Library/Preferences/");
-    #endif
-
-    pathsAddHome(&state->configDirs, state, "");
-
-    #ifdef _WIN32
-        if(getenv("MSYSTEM") && getenv("HOME"))
-        {
-            // We are in MSYS2 / Git Bash
-            pathsAddEnvSuffix(&state->configDirs, "HOME", ".config/");
-            pathsAddEnvSuffix(&state->configDirs, "HOME", "");
-        }
-    #endif
-
-    pathsAddEnv(&state->configDirs, "XDG_CONFIG_DIRS");
-
-    #if !defined(_WIN32) && !defined(__APPLE__)
-        pathsAddAbsolute(&state->configDirs, FASTFETCH_TARGET_DIR_ETC"/xdg/");
-    #endif
-
-    #if !defined(_WIN32)
-        pathsAddAbsolute(&state->configDirs, FASTFETCH_TARGET_DIR_ETC"/");
-        pathsAddAbsolute(&state->configDirs, FASTFETCH_TARGET_DIR_INSTALL_SYSCONF"/");
-    #endif
-}
-
-static void initDataDirs(FFstate* state)
-{
-    ffListInit(&state->dataDirs, sizeof(FFstrbuf));
-
-    pathsAddEnv(&state->dataDirs, "XDG_DATA_HOME"); // On systems where this is not set (Windows & Apple presumably), this wil do nothing
-    pathsAddHome(&state->dataDirs, state, "/.local/share/");
-
-    #ifdef _WIN32
-        pathsAddKnownFolder(&state->dataDirs, &FOLDERID_RoamingAppData);
-        pathsAddKnownFolder(&state->dataDirs, &FOLDERID_LocalAppData);
-    #elif defined(__APPLE__)
-        pathsAddHome(&state->dataDirs, state, "/Library/Application Support/");
-    #endif
-
-    pathsAddHome(&state->dataDirs, state, "");
-
-    #ifdef _WIN32
-        if(getenv("MSYSTEM") && getenv("HOME"))
-        {
-            // We are in MSYS2 / Git Bash
-            pathsAddEnvSuffix(&state->dataDirs, "HOME", ".local/share/");
-            pathsAddEnvSuffix(&state->dataDirs, "HOME", "");
-        }
-    #endif
-
-    pathsAddEnv(&state->dataDirs, "XDG_DATA_DIRS");
-
-    #if !defined(_WIN32)
-        pathsAddAbsolute(&state->dataDirs, FASTFETCH_TARGET_DIR_USR"/local/share/");
-        pathsAddAbsolute(&state->dataDirs, FASTFETCH_TARGET_DIR_USR"/share/");
-    #endif
-}
-
 static void initState(FFstate* state)
 {
-    #ifdef WIN32
-    //https://learn.microsoft.com/en-us/cpp/c-runtime-library/reference/setlocale-wsetlocale?source=recommendations&view=msvc-170#utf-8-support
-    setlocale(LC_ALL, ".UTF8");
-    #endif
-
     state->logoWidth = 0;
     state->logoHeight = 0;
     state->keysHeight = 0;
-    #ifndef WIN32
-        state->passwd = getpwuid(getuid());
-    #else
-        state->passwd = ffGetPasswd();
-    #endif
-    uname(&state->utsname);
 
-    initConfigDirs(state);
-    initDataDirs(state);
+    ffPlatformInit(&state->platform);
 }
 
 static void initModuleArg(FFModuleArgs* args)
@@ -254,7 +83,7 @@ static void defaultConfig(FFinstance* instance)
     initModuleArg(&instance->config.processes);
     initModuleArg(&instance->config.packages);
     initModuleArg(&instance->config.shell);
-    initModuleArg(&instance->config.resolution);
+    initModuleArg(&instance->config.display);
     initModuleArg(&instance->config.de);
     initModuleArg(&instance->config.wm);
     initModuleArg(&instance->config.wmTheme);
@@ -310,6 +139,7 @@ static void defaultConfig(FFinstance* instance)
     ffStrbufInitA(&instance->config.libcJSON, 0);
     ffStrbufInitA(&instance->config.libfreetype, 0);
     ffStrbufInit(&instance->config.libwlanapi);
+    ffStrbufInit(&instance->config.libnm);
 
     instance->config.cpuTemp = false;
     instance->config.gpuTemp = false;
@@ -327,6 +157,7 @@ static void defaultConfig(FFinstance* instance)
     instance->config.diskShowRemovable = true;
     instance->config.diskShowHidden = false;
     instance->config.diskShowUnknown = false;
+    instance->config.diskShowSubvolumes = false;
 
     ffStrbufInitA(&instance->config.batteryDir, 0);
 
@@ -348,10 +179,27 @@ static void defaultConfig(FFinstance* instance)
     ffStrbufInitA(&instance->config.playerName, 0);
 
     instance->config.percentType = 1;
+
+    ffStrbufInitS(&instance->config.commandShell,
+        #ifdef _WIN32
+        "cmd"
+        #elif defined(__FreeBSD__)
+        "csh"
+        #else
+        "bash"
+        #endif
+    );
+    ffListInit(&instance->config.commandKeys, sizeof(FFstrbuf));
+    ffListInit(&instance->config.commandTexts, sizeof(FFstrbuf));
 }
 
 void ffInitInstance(FFinstance* instance)
 {
+    #ifdef WIN32
+        //https://learn.microsoft.com/en-us/cpp/c-runtime-library/reference/setlocale-wsetlocale?source=recommendations&view=msvc-170#utf-8-support
+        setlocale(LC_ALL, ".UTF8");
+    #endif
+
     initState(&instance->state);
     defaultConfig(instance);
 }
@@ -400,6 +248,10 @@ static void resetConsole()
 
     if(ffHideCursor)
         fputs("\033[?25h", stdout);
+
+    #if defined(_WIN32) && defined(FF_ENABLE_BUFFER)
+        fflush(stdout);
+    #endif
 }
 
 #ifdef _WIN32
@@ -493,7 +345,7 @@ static void destroyConfig(FFinstance* instance)
     destroyModuleArg(&instance->config.processes);
     destroyModuleArg(&instance->config.packages);
     destroyModuleArg(&instance->config.shell);
-    destroyModuleArg(&instance->config.resolution);
+    destroyModuleArg(&instance->config.display);
     destroyModuleArg(&instance->config.de);
     destroyModuleArg(&instance->config.wm);
     destroyModuleArg(&instance->config.wmTheme);
@@ -549,6 +401,7 @@ static void destroyConfig(FFinstance* instance)
     ffStrbufDestroy(&instance->config.libcJSON);
     ffStrbufDestroy(&instance->config.libfreetype);
     ffStrbufDestroy(&instance->config.libwlanapi);
+    ffStrbufDestroy(&instance->config.libnm);
 
     ffStrbufDestroy(&instance->config.diskFolders);
     ffStrbufDestroy(&instance->config.batteryDir);
@@ -558,17 +411,19 @@ static void destroyConfig(FFinstance* instance)
     ffStrbufDestroy(&instance->config.weatherOutputFormat);
     ffStrbufDestroy(&instance->config.osFile);
     ffStrbufDestroy(&instance->config.playerName);
+
+    ffStrbufDestroy(&instance->config.commandShell);
+    FF_LIST_FOR_EACH(FFstrbuf, item, instance->config.commandKeys)
+        ffStrbufDestroy(item);
+    ffListDestroy(&instance->config.commandKeys);
+    FF_LIST_FOR_EACH(FFstrbuf, item, instance->config.commandTexts)
+        ffStrbufDestroy(item);
+    ffListDestroy(&instance->config.commandTexts);
 }
 
 static void destroyState(FFinstance* instance)
 {
-    FF_LIST_FOR_EACH(FFstrbuf, configDir, instance->state.configDirs)
-        ffStrbufDestroy(configDir);
-    ffListDestroy(&instance->state.configDirs);
-
-    FF_LIST_FOR_EACH(FFstrbuf, dataDir, instance->state.dataDirs)
-        ffStrbufDestroy(dataDir);
-    ffListDestroy(&instance->state.dataDirs);
+    ffPlatformDestroy(&instance->state.platform);
 }
 
 void ffDestroyInstance(FFinstance* instance)
@@ -648,7 +503,10 @@ void ffListFeatures()
             "libcjson\n"
         #endif
         #ifdef FF_HAVE_FREETYPE
-            "freetype"
+            "freetype\n"
+        #endif
+        #ifdef FF_HAVE_LIBNM
+            "libnm\n"
         #endif
         ""
     , stdout);